Session creating

+++ src/Web/Handlers/QuizHandlers/Delete/DeleteQuizEndpoint.cs
@@ -23,12 +23,12 @@ public class DeleteQuizEndpoint : Endpoint<DeleteQuizRequest, EmptyResponse>
         //if (string.IsNullOrEmpty(userId)) throw new UnauthorizedException();
         //if (await IsQuizTitleTaken(request.Quiz.Title, userId)) throw new QuizTitleTakenException();
         
-        var quiz = await DbContext.Collection<Quiz>().AsQueryable().SingleOrDefaultAsync(q => q.Id.Equals(request.QuizId));
+        var quiz = await DbContext.Collection<Quiz>().AsQueryable().SingleOrDefaultAsync(q => q.Id.Equals(request.QuizId), cancellationToken: cancellationToken);
         if(quiz is null) throw new NotFoundException("Quiz");
 
-        foreach (var question in quiz.Questions.Where(question => question.Image != null))
+        foreach (var question in quiz.Questions)
         {
-            Directory.Delete(question.Image.FullPath, true);
+            if (question.Image != null) Directory.Delete(question.Image.FullPath, true);
         }
 
         await DbContext.Collection<Quiz>().DeleteOneAsync(q => q.Id.Equals(request.QuizId), cancellationToken: cancellationToken);
